requirements.txt
load_dotenv
requests
slack-bolt
openai
Flask>1
slackeventsapi
aiohttp
idna
certifi
urllib3
secrets